[IMP] zen: verses about code certainty
A beautiful code is code that works properly. It is code that works well and its logic is well studied and researched. When in face of ambiguity, don't try to guess what should be done and write code that might fail some cases. Taking time to make sure all points are clear and that the code will behave as intended in all cases will surely save more time that would be taken later in fixing it.